requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
